+++ src/routes/AuthLayoutRouter/AuthLayoutRouter.test.js
@@ -10,6 +10,10 @@ const { ResizeObserver } = window
 jest.mock('mixpanel-browser', () => ({
   track: jest.fn(),
   identify: jest.fn(),
+  get_group: jest.fn().mockImplementation(() => ({
+    set: jest.fn()
+  })),
+  set_group: jest.fn(),
   people: {
     set: jest.fn()
   }